[issue-407] also use add_value_if_exists in creation information
