+++ src/spdx/writer/rdf/creation_info_writer.py
@@ -12,7 +12,7 @@ from rdflib import Graph, BNode, RDF, Literal, RDFS, URIRef
 
 from spdx.datetime_conversions import datetime_to_iso_string
 from spdx.model.document import CreationInfo
-from spdx.writer.rdf.writer_utils import spdx_namespace
+from spdx.writer.rdf.writer_utils import spdx_namespace(), add_literal_value_if_exists
 
 
 def add_creation_info_to_graph(creation_info: CreationInfo, graph: Graph):
@@ -21,6 +21,7 @@ def add_creation_info_to_graph(creation_info: CreationInfo, graph: Graph):
     graph.add((doc_node, spdx_namespace().specVersion, Literal(creation_info.spdx_version)))
     graph.add((doc_node, spdx_namespace().dataLicense, Literal(creation_info.data_license)))
     graph.add((doc_node, spdx_namespace().name, Literal(creation_info.name)))
+    add_literal_value_if_exists(graph, doc_node, RDFS.comment, creation_info.document_comment)
 
     creation_info_node = BNode()
     graph.add((creation_info_node, RDF.type, spdx_namespace().CreationInfo))
@@ -30,11 +31,8 @@ def add_creation_info_to_graph(creation_info: CreationInfo, graph: Graph):
     for creator in creation_info.creators:
         graph.add((creation_info_node, spdx_namespace().creator, Literal(creator.to_serialized_string())))
 
-    graph.add(
-        (creation_info_node, spdx_namespace().licenseListVersion, Literal(str(creation_info.license_list_version))))
-
-    if creation_info.creator_comment:
-        graph.add((creation_info_node, RDFS.comment, Literal(creation_info.creator_comment)))
+    add_literal_value_if_exists(graph, creation_info_node, spdx_namespace().licenseListVersion, creation_info.license_list_version)
+    add_literal_value_if_exists(graph, creation_info_node, RDFS.comment, creation_info.creator_comment)
 
     graph.add((doc_node, spdx_namespace().creationInfo, creation_info_node))
